pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "selfsnap-win11"
version = "0.1.0"
description = "Windows 11 local screenshot capture utility with tray control and Task Scheduler integration."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary - personal project" }
authors = [{ name = "Daniel" }]
dependencies = [
  "mss>=9.0.1",
  "Pillow>=10.3.0",
  "pystray>=0.19.5",
  "pywin32>=306; platform_system == 'Windows'",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.10.0",
  "pyinstaller>=6.6.0",
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.4.0",
  "time-machine>=2.14.1",
]

[project.scripts]
selfsnap = "selfsnap.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "B", "UP", "C4"]

[tool.mypy]
python_version = "3.11"
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
strict_optional = true
disallow_untyped_defs = true
exclude = ["dist", "build"]

